+++ pkg/blackhole/blackhole.go
