+++ pkg/file/file.go
@@ -1,7 +1,7 @@
 package file
 
 import (
-	"github.com/underscorenygren/metrics/pkg/source/stream"
+	"github.com/underscorenygren/metrics/pkg/stream"
 	"github.com/underscorenygren/metrics/pkg/types"
 	"os"
 )
